+++ demo/src/app/components/highlighter/demos/index.ts
@@ -4,11 +4,11 @@ import {NthdHighlighterRegexReplace} from './regex-replace/highlighter-regex-rep
 export const DEMO_DIRECTIVES = [NthdHighlighterBasic, NthdHighlighterRegexReplace];
 export const DEMO_SNIPPETS = {
   'basic': {
-    'code': require('!!raw-loader!./basic/highlighter-basic'),
-    'markup': require('!!raw-loader!./basic/highlighter-basic.html')
+    'code': require('!!raw-loader!./basic/highlighter-basic').default,
+    'markup': require('!!raw-loader!./basic/highlighter-basic.html').default
   },
   'regex-replace': {
-    'code': require('!!raw-loader!./regex-replace/highlighter-regex-replace'),
-    'markup': require('!!raw-loader!./regex-replace/highlighter-regex-replace.html')
+    'code': require('!!raw-loader!./regex-replace/highlighter-regex-replace').default,
+    'markup': require('!!raw-loader!./regex-replace/highlighter-regex-replace.html').default
   }
 };