+++ demo/src/app/components/mentions/demos/index.ts
@@ -14,23 +14,23 @@ export const DEMO_DIRECTIVES = [
 
 export const DEMO_SNIPPETS = {
     'basic': {
-        'code': require('!!raw-loader!./basic/mentions-basic'),
-        'markup': require('!!raw-loader!./basic/mentions-basic.html')
+        'code': require('!!raw-loader!./basic/mentions-basic').default,
+        'markup': require('!!raw-loader!./basic/mentions-basic.html').default
     },
     'custom-search': {
-        'code': require('!!raw-loader!./custom-search/mentions-custom-search'),
-        'markup': require('!!raw-loader!./custom-search/mentions-custom-search.html')
+        'code': require('!!raw-loader!./custom-search/mentions-custom-search').default,
+        'markup': require('!!raw-loader!./custom-search/mentions-custom-search.html').default
     },
     'string-values': {
-        'code': require('!!raw-loader!./string-values/mentions-string-values'),
-        'markup': require('!!raw-loader!./string-values/mentions-string-values.html')
+        'code': require('!!raw-loader!./string-values/mentions-string-values').default,
+        'markup': require('!!raw-loader!./string-values/mentions-string-values.html').default
     },
     'validation': {
-        'code': require('!!raw-loader!./validation/mentions-validation'),
-        'markup': require('!!raw-loader!./validation/mentions-validation.html')
+        'code': require('!!raw-loader!./validation/mentions-validation').default,
+        'markup': require('!!raw-loader!./validation/mentions-validation.html').default
     },
     'custom-template': {
-        'code': require('!!raw-loader!./custom-template/mentions-custom-template'),
-        'markup': require('!!raw-loader!./custom-template/mentions-custom-template.html')
+        'code': require('!!raw-loader!./custom-template/mentions-custom-template').default,
+        'markup': require('!!raw-loader!./custom-template/mentions-custom-template.html').default
     }
 };